Split calendar and calendar action bar
